Process
Handles: The number of handles that the process has opened.
NPM(K): The amount of non-paged memory that the process is using, in kilobytes.
PM(K): The amount of pageable memory that the process is using, in kilobytes.
WS(K): The size of the working set of the process, in kilobytes. The working set consists of the pages of memory that were recently referenced by the process.
VM(M): The amount of virtual memory that the process is using, in megabytes. Virtual memory includes storage in the paging files on disk.
CPU(s): The amount of processor time that the process has used on all processors, in seconds.
ID: The process ID (PID) of the process.
ProcessName: The name of the process. For explanations of the concepts related to processes, see the Glossary in Help and Support Center and the Help for Task Manager.